refactor: Move `DeviceEndpoint` into a separate file
This also moves the bootstrapping connection code into `DeviceEndpoint` as a static factory constructor.
